voicebot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
docopt==0.6.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
langdetect==1.0.9
multidict==6.0.4
openai==0.26.5
pipreqs==0.4.11
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
yarg==0.1.9
yarl==1.8.2
